Renamed block package to workspace
