+++ workspace/metrics_getter.go
@@ -1,4 +1,4 @@
-package block
+package workspace
 
 import (
 	"context"
